# Same implementation for Sparse Multiplication for aligned and unaligned arrays