--- a/test/Microsoft.ML.CpuMath.PerformanceTests/AvxPerformanceTests.cs
+++ b/test/Microsoft.ML.CpuMath.PerformanceTests/AvxPerformanceTests.cs
@@ -108,17 +108,18 @@ namespace Microsoft.ML.CpuMath.PerformanceTests
         [BenchmarkCategory("Fma")]
         public void SdcaL1UpdateSU()
             => AvxIntrinsics.SdcaL1UpdateSU(DefaultScale, IndexLength, src, idx, DefaultScale, dst, result);
+
         [Benchmark]
         [BenchmarkCategory("Fma")]
-        public void MatMulX()
-            => AvxIntrinsics.MatMulX(src, src1, dst, 1000, 1000);
+        public void MatMul()
+            => AvxIntrinsics.MatMul(src, src1, dst, 1000, 1000);
 
         [Benchmark]
-        public void MatMulTranX()
-            => AvxIntrinsics.MatMulTranX(src, src1, dst, 1000, 1000);
+        public void MatMulTran()
+            => AvxIntrinsics.MatMulTran(src, src1, dst, 1000, 1000);
 
         [Benchmark]
-        public void MatMulPX()
-            => AvxIntrinsics.MatMulPX(src, idx, src1, 0, 0, 8, dst, 1000, 1000);
+        public void MatMulP()
+            => AvxIntrinsics.MatMulP(src, idx, src1, 0, 0, 8, dst, 1000, 1000);
     }
 }