… src/Microsoft.ML.CpuMath/Avx.cs
… src/Microsoft.ML.CpuMath/SseIntrinsics.cs
… src/Native/CpuMathNative/Avx.cpp
… src/Native/CpuMathNative/CMakeLists.txt
… test/Microsoft.ML.CpuMath.PerformanceTests/PerformanceTests.cs
… test/Microsoft.ML.CpuMath.UnitTests.netcoreapp/UnitTests.cs