Working towards https://github.com/dotnet/machinelearning/issues/1018