@eerhardt 2018-10-18
> #endif

It looks like this whole block can be removed as well. It is no longer used.

---
Refers to: src/Native/CpuMathNative/Sse.cpp:79 in db669af. [](commit_id = db669afaf5d8f44aa3a7361a952dbee67f6b8afb, deletion_comment = False)