@Anipik 2018-10-22
``` ini
Before
```
|      Method |     Avx  |     Sse   |    Native |
|------------ |---------:|----------:|----------:|
| MatMulPX(S = 10%)     | 95.28 us | 114.00 us | 91.8  us |
| MatMulPX (S = 20%)| 146.2 us | 181.2 us | 131.3  us |


``` ini
After
```
|      Method |     Avx  |     Sse   |    Native |
|------------ |---------:|----------:|----------:|
| MatMulPX(S = 10%)     | 96.46 us | 106.3 us | 92.8  us |
| MatMulPX (S = 20%)| 140 us | 171.7 us | 131.1  us |


As the matrix becomes more dense , the new algorithm becomes faster

cc @danmosemsft @eerhardt @tannergooding