@Anipik 2018-10-23
@tannergooding i have resolved the conflicts and addressed the feedback