@Anipik 2018-10-23
I have restarted the queue and new build passed successfully